requirements.txt
pyyaml~=6.0
pydantic~=1.9.0
uvicorn~=0.17.1
fastapi~=0.73.0